adds exp stuff for the player
