+++ Assets/Scenes/SampleScene.unity
@@ -2067,7 +2067,7 @@ GameObject:
   - component: {fileID: 1977409044}
   m_Layer: 5
   m_Name: UpgradeMenu
-  m_TagString: Untagged
+  m_TagString: UpgradeMenu
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
