+++ Assets/Scripts/Scriptable Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a4b4fd16180409c498f904a3ac18ba04
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
